fixed time adjustment in realign_spikes
This was broken previously, but there wasn't a test for it. Times needed to be
adjusted by the shift in the peak.

Breaking change for the expected_peak parameter; it's now corrected for
upsampling so that the user can provide the expected peak in the un-resampled
signal (which is more intuitive)